requirements.txt
telethon
requests
bs4
python-dotenv